refactor(move): 설정 파일 디렉토리 이동
- web, security 디렉토리로 나누어 구분
