+++ src/main/java/kr/co/automl/domain/user/controller/UserController.java
@@ -2,7 +2,7 @@ package kr.co.automl.domain.user.controller;
 
 import kr.co.automl.domain.user.dto.SessionUser;
 import kr.co.automl.domain.user.dto.UserInfo;
-import kr.co.automl.global.config.LoginUser;
+import kr.co.automl.global.config.web.LoginUser;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
