+++ src/main/java/kr/co/automl/global/config/web/LoginUser.java
@@ -1,4 +1,4 @@
-package kr.co.automl.global.config;
+package kr.co.automl.global.config.web;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
