+++ src/main/java/kr/co/automl/global/config/web/WebConfig.java
@@ -1,4 +1,4 @@
-package kr.co.automl.global.config;
+package kr.co.automl.global.config.web;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
