+++ src/test/java/kr/co/automl/domain/user/controller/UserControllerTest.java
@@ -3,8 +3,8 @@ package kr.co.automl.domain.user.controller;
 import kr.co.automl.domain.user.Role;
 import kr.co.automl.domain.user.dto.SessionUser;
 import kr.co.automl.domain.user.dto.UserInfo;
-import kr.co.automl.global.config.LoginUser;
-import kr.co.automl.global.config.LoginUserArgumentResolver;
+import kr.co.automl.global.config.web.LoginUser;
+import kr.co.automl.global.config.web.LoginUserArgumentResolver;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
